+++ packages/piveau-hub-ui-modules/lib/data-provider-interface/config/dcatap/input-definition.ts
@@ -6,49 +6,49 @@ import config from './page-content-config';
 /**
  * Available properties for datasets.
  */
-export type DcatApDatasetsProperty = 
+export type DcatApDatasetsProperty =
   // Append new properties here for accurate type checking
-  'datasetID' 
+  'datasetID'
   | 'description'
   | 'title'
   | 'contactPoint'
   | 'subject'
-  | 'keyword' 
-  | 'publisher' 
-  | 'spatial' 
-  | 'temporal' 
-  | 'theme' 
-  | 'accessRights' 
-  | 'creator' 
-  | 'conformsTo' 
-  | 'page' 
-  | 'accrualPeriodicity' 
-  | 'hasVersion' 
-  | 'isVersionOf' 
-  | 'source' 
-  | 'identifier' 
-  | 'isReferencedBy' 
-  | 'landingPage' 
-  | 'language' 
-  | 'admsIdentifier' 
-  | 'provenance' 
-  | 'qualifiedAttribution' 
-  | 'wasGeneratedBy' 
-  | 'qualifiedRelation' 
-  | 'relation' 
-  | 'issued' 
-  | 'modified' 
-  | 'spatialResolutionInMeters' 
-  | 'temporalResolution' 
-  | 'type' 
-  | 'versionInfo' 
-  | 'versionNotes' 
-  | 'catalog' 
+  | 'keyword'
+  | 'publisher'
+  | 'spatial'
+  | 'temporal'
+  | 'theme'
+  | 'accessRights'
+  | 'creator'
+  | 'conformsTo'
+  | 'page'
+  | 'accrualPeriodicity'
+  | 'hasVersion'
+  | 'isVersionOf'
+  | 'source'
+  | 'identifier'
+  | 'isReferencedBy'
+  | 'landingPage'
+  | 'language'
+  | 'admsIdentifier'
+  | 'provenance'
+  | 'qualifiedAttribution'
+  | 'wasGeneratedBy'
+  | 'qualifiedRelation'
+  | 'relation'
+  | 'issued'
+  | 'modified'
+  | 'spatialResolutionInMeters'
+  | 'temporalResolution'
+  | 'type'
+  | 'versionInfo'
+  | 'versionNotes'
+  | 'catalog'
   | 'isUsedBy';
 
-  /**
-   * Available properties for distributions.
-   */
+/**
+ * Available properties for distributions.
+ */
 export type DcatApDistributionsProperty =
   // Append new properties here for accurate type checking
   'accessURL'
@@ -80,12 +80,15 @@ export type DcatApDistributionsProperty =
   | 'temporalResolution'
   | 'type'
   | 'status';
-  
+
 export type InputDefinition = {
   datasets: Record<DcatApDatasetsProperty, FormKitSchemaDefinition>;
   distributions: Record<DcatApDistributionsProperty, FormKitSchemaDefinition>;
   // todo complete catalogs type definition
   catalogues: Record<string, FormKitSchemaDefinition>;
+}
+let sectionSchemaTooltips = {
+  prefix: { $el: 'div', attrs: { class: 'infoI', }, children: [{ $el: 'div', children: 'Pass TooltipText to Schema?', attrs: { class: 'tooltipFormkit' } }] }
 }
 
 const dcatapProperties: InputDefinition = {
@@ -114,22 +117,24 @@ const dcatapProperties: InputDefinition = {
               value: 'en',
               $formkit: 'select',
               options: language,
-              'prefix-icon': "info",
+
               validation: 'required',
               name: '@language',
               classes: {
                 outer: 'w25-textfield'
-              }
+              },
+              'sections-schema': sectionSchemaTooltips
             },
             {
               identifier: 'description',
               $formkit: 'textarea',
-              'prefix-icon': "info",
+
               name: '@value',
               validation: 'required',
               classes: {
                 outer: 'w75-descField'
-              }
+              },
+              'sections-schema': sectionSchemaTooltips
             },
           ],
         }
@@ -137,7 +142,7 @@ const dcatapProperties: InputDefinition = {
     },
     // minimum (unability to remove fields to 0)
     title: {
-      identifier: 'title',
+      id: 'title',
       $formkit: 'repeatable',
       name: 'dct:title',
       children: [
@@ -154,21 +159,23 @@ const dcatapProperties: InputDefinition = {
               $formkit: 'select',
               validation: 'required',
               options: language,
-              'prefix-icon': "info",
+
               name: '@language',
               classes: {
                 outer: 'w25-textfield'
-              }
+              },
+              'sections-schema': sectionSchemaTooltips
             },
             {
               identifier: 'titleLabel',
               $formkit: 'text',
               name: '@value',
               validation: 'required',
-              'prefix-icon': "info",
+
               classes: {
                 outer: 'w75-textfield'
-              }
+              },
+              'sections-schema': sectionSchemaTooltips
             },
           ],
         }
@@ -288,14 +295,16 @@ const dcatapProperties: InputDefinition = {
     subject: {
       identifier: 'subject',
       $formkit: 'auto',
-      name: 'dct:subject',
+      property: 'dct:subject',
       multiple: true,
       annifTheme: true,
       voc: 'eurovoc',
+      // name:'dct:subject',
       '@annifSuggestion': false,
       classes: {
-        outer: 'w88-textfield'
+        outer: 'w97-textfield'
       },
+      'sections-schema': sectionSchemaTooltips
     },
     keyword: {
       identifier: 'keyword',
@@ -332,8 +341,10 @@ const dcatapProperties: InputDefinition = {
     publisher: {
       $formkit: 'auto',
       identifier: 'publisher',
-      name: 'dct:publisher',
       voc: 'corporate-body',
+      property: 'dct:publisher',
+      id: 'publisher',
+      name: 'dct:publisher',
     },
     spatial: {
       identifier: 'spatial',
@@ -440,12 +451,15 @@ const dcatapProperties: InputDefinition = {
       multiple: true,
       annifTheme: true,
       voc: 'data-theme',
-      name: 'dcat:theme',
       '@annifSuggestion': false,
+      name: 'dcat:theme',
+      id: 'theme',
       classes: {
         multiple: 'multiple',
         outer: 'w88-textfield'
       },
+
+
     },
     accessRights: {
       identifier: 'accessRights',
@@ -596,8 +610,8 @@ const dcatapProperties: InputDefinition = {
               $formkit: 'auto',
               identifier: 'pageFormat',
               voc: 'file-type',
-              name: 'dct:format',
               class: "property",
+              name: 'dct:format',
             },
             {
               identifier: 'pageUrl',
@@ -618,6 +632,7 @@ const dcatapProperties: InputDefinition = {
       $formkit: 'auto',
       voc: 'frequency',
       name: 'dct:accrualPeriodicity',
+
     },
     hasVersion: {
       identifier: 'hasVersion',
@@ -924,7 +939,7 @@ const dcatapProperties: InputDefinition = {
         {
           identifier: 'issued',
           id: 'issuedCond',
-          'prefix-icon': "info",
+
           $formkit: 'select',
           name: '@type',
           options: { date: 'Date', datetime: 'Datetime' },
@@ -943,7 +958,8 @@ const dcatapProperties: InputDefinition = {
               type: 'datetime-local',
               name: '@value',
             }
-          }
+          },
+          'sections-schema': sectionSchemaTooltips
         },
       ]
     },
@@ -956,7 +972,7 @@ const dcatapProperties: InputDefinition = {
           identifier: 'modified',
           id: 'modifiedCond',
           name: '@type',
-          'prefix-icon': "info",
+
           $formkit: 'select',
           options: { date: 'Date', datetime: 'Datetime' },
         },
@@ -975,7 +991,8 @@ const dcatapProperties: InputDefinition = {
               type: 'datetime-local',
               name: '@value'
             }
-          }
+          },
+          'sections-schema': sectionSchemaTooltips
         },
       ]
     },
@@ -1075,9 +1092,10 @@ const dcatapProperties: InputDefinition = {
       identifier: 'catalog',
       $formkit: 'select',
       name: 'dcat:catalog',
-      'prefix-icon': "info",
+
       validation: 'required',
-      options: {},
+      options: { "title": "hallo" },
+      'sections-schema': sectionSchemaTooltips
     },
     isUsedBy: {
       identifier: 'isUsedBy',
@@ -1571,7 +1589,8 @@ const dcatapProperties: InputDefinition = {
               type: 'datetime-local',
               name: '@value',
             }
-          }
+          },
+          'sections-schema': sectionSchemaTooltips
         },
       ]
     },
@@ -1601,7 +1620,8 @@ const dcatapProperties: InputDefinition = {
               type: 'datetime-local',
               name: '@value'
             }
-          }
+          },
+          'sections-schema': sectionSchemaTooltips
         },
       ]
     },
@@ -1701,7 +1721,7 @@ const dcatapProperties: InputDefinition = {
   catalogues: {
     datasetID: {
       // todo: check if this is correct
-      $formkit:'id',
+      $formkit: 'id',
       identifier: 'datasetID',
       name: 'datasetID',
       mandatory: true,
@@ -2018,6 +2038,7 @@ const dcatapProperties: InputDefinition = {
     },
   }
 };
+
 
 // Dynamically add a collapsed property to all fields that are component of
 // a set of specific pages steps.